debian: change libssl dependency
Depending on libssl 1.0 is not a good idea: as we can see on their release
strategy map[1], version 1.0.2 will (only) be supported until 2019-12-31 (LTS).

That mas openssl 1.0 an unsuitable target for debian and ubuntu's latest
versions, that no longer have that library in their repositories, making this
package unusable for users of the latest debian or ubuntu versions, even now.

[1] https://www.openssl.org/policies/releasestrat.html